refactor(study): study controller to api endpoints
refactor study controller to api endpoints

Closes #811
