+++ src/Sepes.RestApi/ApiEndpoints/Studies/Details.cs
@@ -2,29 +2,32 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Sepes.Infrastructure.Service.Interface;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
-namespace Sepes.RestApi.Controller
+namespace Sepes.RestApi.ApiEndpoints.Studies
 {
     [Route("api/studies")]
     [ApiController]
     [Produces("application/json")]
     [EnableCors("_myAllowSpecificOrigins")]
     [Authorize]
-    public class StudyDetailsController : ControllerBase
+    public class Details : ControllerBase
     {
-        readonly IStudyDetailsService _studyDetailsService;        
+        readonly IStudyDetailsService _studyDetailsService;
 
-        public StudyDetailsController(IStudyDetailsService studyDetailsService)
+        public Details(IStudyDetailsService studyDetailsService)
         {
-            _studyDetailsService = studyDetailsService;             
-        }      
+            _studyDetailsService = studyDetailsService;
+        }
 
         [HttpGet("{studyId}")]
-        public async Task<IActionResult> GetAsync(int studyId)
+        public async Task<IActionResult> Handle(int studyId)
         {
             var study = await _studyDetailsService.Get(studyId);
             return new JsonResult(study);
-        }                    
+        }
     }
 }
