+++ src/Sepes.RestApi/ApiEndpoints/Studies/List.cs
@@ -2,9 +2,12 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Sepes.Infrastructure.Service.Interface;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
-namespace Sepes.RestApi.Controller
+namespace Sepes.RestApi.ApiEndpoints.Studies
 {
     [Route("api/studies")]
     [ApiController]
@@ -14,20 +17,16 @@ namespace Sepes.RestApi.Controller
     public class StudyListController : ControllerBase
     {
         readonly IStudyListService _studyListService;
-     
-
         public StudyListController(IStudyListService studyListService)
         {
             _studyListService = studyListService;
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetStudiesAsync()
+        public async Task<IActionResult> Handle()
         {
             var studies = await _studyListService.GetStudyListAsync();
             return new JsonResult(studies);
         }
-
-           
     }
 }
